requirements.txt
SQLAlchemy
mysqlclient
paho-mqtt
Flask
influxdb